[pankaj]added jms and token to user

+++ src/main/java/com/BridgeLabz/BookstoreApp/entity/User.java
@@ -26,51 +26,51 @@ public class User {
     public User() {
     }
 
-    public User(UserDTO userDTO){
-        this.id=getId();
-        this.firstname= userDTO.getFirstname();
-        this.lastname= userDTO.getLastname();
-        this.kyc= userDTO.getKyc();
-        this.dob= userDTO.getDob();
-        this.registeredDate= userDTO.getRegisteredDate();
-        this.updatedDate=userDTO.getUpdatedDate();
-        this.email=userDTO.getEmail();
-        this.password=userDTO.getPassword();
+    public User(UserDTO userDTO) {
+        this.id = getId();
+        this.firstname = userDTO.getFirstname();
+        this.lastname = userDTO.getLastname();
+        this.kyc = userDTO.getKyc();
+        this.dob = userDTO.getDob();
+        this.registeredDate = userDTO.getRegisteredDate();
+        this.updatedDate = userDTO.getUpdatedDate();
+        this.email = userDTO.getEmail();
+        this.password = userDTO.getPassword();
     }
 
-    public User(Long id,UserDTO userDTO){
-        this.id=id;
-        this.firstname= userDTO.getFirstname();
-        this.lastname= userDTO.getLastname();
-        this.kyc= userDTO.getKyc();
-        this.dob= userDTO.getDob();
-        this.registeredDate= userDTO.getRegisteredDate();
-        this.updatedDate=userDTO.getUpdatedDate();
-        this.email=userDTO.getEmail();
-        this.password=userDTO.getPassword();
+    public User(Integer id, UserDTO userDTO) {
+        this.id = id;
+        this.firstname = userDTO.getFirstname();
+        this.lastname = userDTO.getLastname();
+        this.kyc = userDTO.getKyc();
+        this.dob = userDTO.getDob();
+        this.registeredDate = userDTO.getRegisteredDate();
+        this.updatedDate = userDTO.getUpdatedDate();
+        this.email = userDTO.getEmail();
+        this.password = userDTO.getPassword();
     }
 
-    public User(String email_id,UserDTO userDTO){
-        this.email=email_id;
-        this.firstname= userDTO.getFirstname();
-        this.lastname= userDTO.getLastname();
-        this.kyc= userDTO.getKyc();
-        this.dob= userDTO.getDob();
-        this.registeredDate= userDTO.getRegisteredDate();
-        this.updatedDate=userDTO.getUpdatedDate();
-        this.password=userDTO.getPassword();
+    public User(String email_id, UserDTO userDTO) {
+        this.email = email_id;
+        this.firstname = userDTO.getFirstname();
+        this.lastname = userDTO.getLastname();
+        this.kyc = userDTO.getKyc();
+        this.dob = userDTO.getDob();
+        this.registeredDate = userDTO.getRegisteredDate();
+        this.updatedDate = userDTO.getUpdatedDate();
+        this.password = userDTO.getPassword();
     }
 
-    public void updateUser(UserDTO user){
+    public void updateUser(UserDTO user) {
 
-        this.firstname= user.getFirstname();
-        this.lastname= user.getLastname();
-        this.kyc= user.getKyc();
-        this.dob= user.getDob();
-        this.registeredDate= user.getRegisteredDate();
-        this.updatedDate=user.getUpdatedDate();
-        this.email=user.getEmail();
-        this.password=user.getPassword();
+        this.firstname = user.getFirstname();
+        this.lastname = user.getLastname();
+        this.kyc = user.getKyc();
+        this.dob = user.getDob();
+        this.registeredDate = user.getRegisteredDate();
+        this.updatedDate = user.getUpdatedDate();
+        this.email = user.getEmail();
+        this.password = user.getPassword();
     }
